# swagger.yaml
openapi: 3.0.3
info:
  title: Swagger API Course FE
  description: |-
    This is a list api of course FE
  contact:
    email: botono123@gmail.com
  version: 1.0.11
servers:
  - url: http://localhost:3001
...
paths:
...
  /api/roles:
    post:
      tags:
        - roles
      summary: Add a new role
      description: Add a new role
      operationId: addRole
      requestBody:
        description: Create a new role
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Role"
          application/xml:
            schema:
              $ref: "#/components/schemas/Role"
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/Role"
        required: true
      responses:
        "201":
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                items:
                  $ref: "#/components/schemas/Role"
            application/xml:
              schema:
                type: object
                items:
                  $ref: "#/components/schemas/Role"
        "401":
          description: UNAUTHORIZED
        "500":
          description: INTERNAL_SERVER_ERROR
        "409":
          description: ALREADY_EXIST
        "400":
          description: INVALID
      security:
        - bearerAuth: []
    get:
      tags:
        - roles
      summary: Get all role
      description: get all roles
      operationId: getAllRole
      parameters:
        - name: limit
          in: query
          description: Limit of list roles
          required: false
          example: 10
          schema:
            type: number
        - name: page
          in: query
          description: Page of list roles
          required: false
          example: 1
          schema:
            type: number
        - name: search
          in: query
          description: search of list roles
          required: false
          schema:
            type: string
        - name: order
          in: query
          description: sort of list roles
          required: false
          example: "created asc"
          schema:
            type: string
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Role"
            application/xml:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Role"
        "401":
          description: UNAUTHORIZED
        "500":
          description: INTERNAL_SERVER_ERROR
        "409":
          description: ALREADY_EXIST
        "400":
          description: INVALID
      security:
        - bearerAuth: []
  /api/roles/{roleId}:
    get:
      tags:
        - roles
      summary: Get details roles order by ID
      description: Get details role by id of roles
      operationId: getRoleById
      parameters:
        - name: roleId
          in: path
          description: ID of role that needs to be fetched
          required: true
          schema:
            type: string
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                items:
                  $ref: "#/components/schemas/Role"
            application/xml:
              schema:
                type: object
                items:
                  $ref: "#/components/schemas/Role"
        "401":
          description: UNAUTHORIZED
        "500":
          description: INTERNAL_SERVER_ERROR
        "409":
          description: ALREADY_EXIST
        "400":
          description: INVALID
    delete:
      tags:
        - roles
      summary: Delete rOLE by ID
      description: Try delete user by id of user
      operationId: deleteUser
      parameters:
        - name: userId
          in: path
          description: ID of the user that needs to be deleted
          required: true
          schema:
            type: string
      responses:
        "201":
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                items:
                  $ref: "#/components/schemas/Role"
            application/xml:
              schema:
                type: object
                items:
                  $ref: "#/components/schemas/Role"
        "401":
          description: UNAUTHORIZED
        "500":
          description: INTERNAL_SERVER_ERROR
        "409":
          description: ALREADY_EXIST
        "400":
          description: INVALID
    put:
      tags:
        - roles
      summary: Update an existing role
      description: Update an existing role by Id
      operationId: updateRole
      parameters:
        - name: roleId
          in: path
          description: role id of role updated
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                permissions:
                  type: array
                  format: string
                  example: []
              required:
                - name
                - permissions
          application/xml:
            schema:
              type: object
              properties:
                name:
                  type: string
                permissions:
                  type: array
                  format: string
              required:
                - name
                - permissions
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                name:
                  type: string
                permissions:
                  type: array
                  format: string
              required:
                - name
                - permissions

      responses:
        "201":
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                items:
                  $ref: "#/components/schemas/Role"
            application/xml:
              schema:
                type: object
                items:
                  $ref: "#/components/schemas/Role"
        "401":
          description: UNAUTHORIZED
        "500":
          description: INTERNAL_SERVER_ERROR
        "409":
          description: ALREADY_EXIST
        "400":
          description: INVALID
      security:
        - bearerAuth: []
  /api/roles/delete-many:
    delete:
      tags:
        - roles
      summary: Delete many roles by ids
      description: Delete many roles by list ids
      operationId: deleteManyRole
      requestBody:
        description: Object chứa mảng id của role cần xóa
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                roleIds:
                  type: array
                  items:
                    type: string
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Role"
            application/xml:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Role"
        "201":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Role"
            application/xml:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Role"
        "401":
          description: UNAUTHORIZED
        "500":
          description: INTERNAL_SERVER_ERROR
        "409":
          description: ALREADY_EXIST
        "400":
          description: INVALID
      security:
        - bearerAuth: []
...
components:
  schemas:
...
    Role:
      type: object
      properties:
        name:
          type: string
        permissions:
          type: array
      required:
        - name
        - permissions
      xml:
        name: Role
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT